+++ src/cli/templates/list-template.js
@@ -9,10 +9,10 @@ import { BitId } from '../../bit-id';
 export default (components: ConsumerComponent[], json: boolean, showRemoteVersion?: boolean = false) => {
   const header = [
     { value: 'Id', width: 70, headerColor: 'cyan', headerAlign: 'left' },
-    { value: showRemoteVersion ? 'Local Version' : 'Version', width: 9, headerColor: 'cyan', headerAlign: 'left' }
+    { value: showRemoteVersion ? 'Local Version' : 'Version', width: 9, headerColor: 'cyan', headerAlign: 'left' },
+    { value: 'Current Version', width: 9, headerColor: 'cyan', headerAlign: 'left' }
   ];
   if (showRemoteVersion) {
-    header.push({ value: 'Current Version', width: 9, headerColor: 'cyan', headerAlign: 'left' });
     header.push({ value: 'Remote Version', width: 9, headerColor: 'cyan', headerAlign: 'left' });
   }
   const opts = {
@@ -34,8 +34,8 @@ export default (components: ConsumerComponent[], json: boolean, showRemoteVersio
       return bitId.version;
     };
     data.localVersion = version;
+    data.currentVersion = getCurrentlyUsedVersion();
     if (showRemoteVersion) {
-      data.currentVersion = getCurrentlyUsedVersion();
       let remoteVersion = component.latest || 'N/A';
       const color = component.latest && semver.gt(component.version, component.latest) ? 'red' : null;
       remoteVersion = color ? c[color](remoteVersion) : remoteVersion;